requirements.txt
tensorflow-gpu==2.0.1
numpy==1.17.4
opencv-python
Pillow
tqdm
matplotlib
imgaug
